requirements.txt
sentence-transformers
scikit-learn
spacy
pypdf
networkx
python-louvain
ollama
langchain
nltk
numpy
pyyaml
typer
rich
tiktoken
pytest
